Refactor single path integration test to it's own file
